FATAL: odometry: change the robot speed

+++ src/blackmetal/src/communication/Client.hpp
@@ -61,7 +61,7 @@ public:
 	 * 		   bm::Status::RECEIVE_ERROR when the ::read function crashes,
 	 * 		   returned std::string message otherwise.
 	 */
-	virtual std::variant<bm::Status, std::string> execute(bm::Command cmd, int rightWheel = 0, int leftWheel = 0);
+	virtual std::variant<bm::Status, std::string> execute(bm::Command cmd, double rightWheel = 0, double leftWheel = 0);
 
 	/**
 	 * @brief A specific function just for calling execute with bm::Command::SET_LR_WHEEL_VELOCITY.
@@ -74,7 +74,7 @@ public:
 	 * 		   bm::Status::RECEIVE_ERROR when the ::read function crashes,
 	 * 		   bm::Status::OK otherwise.
 	 */
-	bm::Status request(int rightWheel, int leftWheel);
+	bm::Status request(double rightWheel, double leftWheel);
 
 	/**
 	 * @brief Send a desired string message to the server.
